# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: manageRTC HRM API
  description: |
    Human Resource Management API for the manageRTC application.

    Features include:
    - Timesheet management with approval workflow
    - Leave request management
    - Overtime request handling
    - Employee lifecycle management (promotions, resignations, terminations)
    - Attendance tracking
    - Department and designation management

    **Authentication**: All endpoints require Bearer token authentication via Clerk JWT.

    **Multi-Tenant**: The API uses company-specific database collections. Include `X-Company-Id` header or use JWT claim.
  version: 1.0.0
  contact:
    name: manageRTC Support
    email: support@manageRTC.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.managertc.com/api
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /overtime:
    get:
      tags: [Overtime]
      summary: Get overtime requests
      description: Retrieve paginated overtime requests.
      operationId: getOvertimeRequests
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, cancelled]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OvertimeRequest'
  /overtime/{id}:
    put:
      tags: [Overtime]
      summary: Update overtime request
      description: Update an existing overtime request.
      operationId: updateOvertimeRequest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: string
                  format: date
                hours:
                  type: number
                  minimum: 0.5
                  maximum: 12
                reason:
                  type: string
      responses:
        '200':
          description: Overtime request updated successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token from Clerk authentication.

        Include in Authorization header:
        ```
        Authorization: Bearer <token>
        ```
...
  schemas:
...
    OvertimeRequest:
      type: object
      properties:
        _id:
          type: string
          format: uuid
        employeeId:
          type: string
          format: uuid
        employeeName:
          type: string
          readOnly: true
        date:
          type: string
          format: date
        hours:
          type: number
          minimum: 0.5
          maximum: 12
        reason:
          type: string
          minLength: 10
        status:
          type: string
          enum: [pending, approved, rejected, cancelled]
        createdAt:
          type: string
          format: date-time
          readOnly: true
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: 'Validation failed'
        message:
          type: string
          example: 'Entries are required'
        details:
          type: object
          additionalProperties: true
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: 'Validation Error'
            message: 'weekStart is required'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: 'Not Found'
            message: 'Timesheet not found'